requirements.txt
flask==2.2.5
flask-cors==3.0.10
python-dotenv==0.21.1
scikit-learn==1.2.2
numpy==1.24.3
pandas==1.5.3
gunicorn==20.1.0
flask-restx==1.0.6
flask-sqlalchemy==3.0.5
flask-jwt-extended==4.5.2
bcrypt==4.0.1
sqlalchemy==2.0.20 